Fix/endpoints-ibrahim

+++ DeciBel/src/main/java/software/decibel/services/track/TrackService.java
@@ -1,12 +1,15 @@
 package software.decibel.services.track;
 
+import jakarta.transaction.Transactional;
 import java.io.IOException;
 import java.time.Duration;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
-
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.text.WordUtils;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -17,17 +20,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.support.TransactionSynchronization;
 import org.springframework.transaction.support.TransactionSynchronizationManager;
 import org.springframework.web.server.ResponseStatusException;
-
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
+import software.decibel.dtos.Resource;
 import software.decibel.dtos.track.requests.TrackPatchRequest;
 import software.decibel.dtos.track.requests.TrackUploadRequest;
 import software.decibel.dtos.track.responses.TrackPageResponse;
 import software.decibel.dtos.track.responses.TrackPatchResponse;
 import software.decibel.dtos.track.responses.TrackPublishResponse;
 import software.decibel.dtos.track.responses.TrackResponse;
-import software.decibel.dtos.Resource;
 import software.decibel.dtos.track.responses.TrackStatusResponse;
 import software.decibel.dtos.track.responses.TrackUploadResponse;
 import software.decibel.dtos.track.responses.TrackWaveFormUrlResponse;
@@ -148,7 +147,13 @@ public class TrackService {
     //delete track
     @Transactional
     public void deleteTrack(Long trackId) {
-        Track track = trackChecksUtil.getTrackIfExistsById(trackId);
+    Track track = trackChecksUtil.getTrackIfExistsById(trackId);
+
+    User uploader = track.getUploader();
+    Long currentUserId = JwtService.getCurrentUserId();
+    if (!Objects.equals(currentUserId, uploader.getId())) {
+      throw new UnauthorizedActionException("You cannot delete a track you didn't upload.");
+    }
 
         //fetch track url data before deleting
         final String audioUrl = track.getTrackUrl();
@@ -158,13 +163,13 @@ public class TrackService {
         likeRepository.deleteAllByTrackId(trackId);
         repostRepository.deleteAllByTrackId(trackId);
         commentRepository.deleteAllByTrackId(trackId);
-        User user = track.getUploader();
-        //update track count
-        user.setTrackCount(user.getTrackCount() - 1);
 
-        // if free user deleted a nn-blocked track they free a slot
-        if (track.getAccess() != TrackAccess.BLOCKED && user.getTier() == AccountTier.FREE) {
-            user.setFreeTracksLeft(user.getFreeTracksLeft() + 1);
+    // update track count
+    uploader.setTrackCount(uploader.getTrackCount() - 1);
+
+    // if free uploader deleted a nn-blocked track they free a slot
+    if (track.getAccess() != TrackAccess.BLOCKED && uploader.getTier() == AccountTier.FREE) {
+      uploader.setFreeTracksLeft(uploader.getFreeTracksLeft() + 1);
         }
 
         trackRepository.delete(track);
@@ -315,6 +320,13 @@ public class TrackService {
     @Transactional
     public void deleteTrackCover(Long trackId) {
         Track track = trackChecksUtil.getTrackIfExistsById(trackId);
+
+    User uploader = track.getUploader();
+    Long currentUserId = JwtService.getCurrentUserId();
+    if (!Objects.equals(currentUserId, uploader.getId())) {
+      throw new UnauthorizedActionException(
+          "You cannot delete the cover for track you didn't upload.");
+    }
         if (track.getCoverUrl() != null) {
             fileUtilityAzure.deleteFileByUrl(track.getCoverUrl());
             track.setCoverUrl(null);
@@ -351,11 +363,16 @@ public class TrackService {
 
     @Transactional
     public TrackPatchResponse updateTrack(Long trackId, TrackPatchRequest request) {
-        Track track = trackChecksUtil.getTrackIfExistsById(trackId);
-        Long userId = JwtService.getCurrentUserId();
-        User uploader = userService.getUserIfExistsById(userId);
 
-        if (request.title() != null) {
+    Track track = trackChecksUtil.getTrackIfExistsById(trackId);
+        Long userId = JwtService.getCurrentUserId();
+    User user = userService.getUserIfExistsById(userId);
+
+    // only uploader can patch
+    if (!Objects.equals(userId, track.getUploader().getId())) {
+      throw new UnauthorizedActionException("You cannot update a track you didn't upload.");
+    }
+    if (request.title() != null) {
             track.setTitle(request.title());
         }
         if (request.genre() != null) {
@@ -384,11 +401,11 @@ public class TrackService {
         }
 
         if (request.access() != null) {
-            TrackAccess finalAccess
-                    = trackPlaybackService.resolvePatchAccess(uploader, track.getAccess(), request.access());
+      TrackAccess finalAccess =
+          trackPlaybackService.resolvePatchAccess(user, track.getAccess(), request.access());
 
-            // update free tracks left based on initial access and final access
-            trackPlaybackService.updateFreeTracksLeft(uploader, track.getAccess(), finalAccess);
+      // update free tracks left based on initial access and final access
+      trackPlaybackService.updateFreeTracksLeft(user, track.getAccess(), finalAccess);
             track.setAccess(finalAccess);
         }
 
@@ -439,7 +456,8 @@ public class TrackService {
         Track track = trackChecksUtil.getTrackIfExistsById(trackId);
 
         if (!track.getUploader().getId().equals(JwtService.getCurrentUserId())) {
-            throw new UnauthorizedActionException("You are not allowed to publish this track.");
+      throw new UnauthorizedActionException(
+          "You are not allowed to publish a track you didn't upload.");
         }
 
         if (track.isPublished()) {
